cleaned up how it works home page
