+++ src/components/pages/CausePage/CauseHomePage.js
@@ -4,7 +4,6 @@ import {getContent} from "./../../../actions"
 import DonateSection from "./sections/DonateSection";
 import OurDonorSection from "./sections/OurDonorSection";
 import JoinSection from "./sections/JoinSection";
-import {Loading} from '../../modules';
 import FirstSection from '../CausePage/sections/FirstSection';
 
 
@@ -13,7 +12,6 @@ class causesHomePage extends Component {
 
     render () {
         const { content, isFetching, error} = this.props;
-        console.log("this", content);
 
         return (
             <div>
