+++ src/components/pages/HowItWorksPage/sections/FirstSection.js
@@ -25,15 +25,22 @@ class FirstSection extends Component {
             return <div>error</div>
         }
 
-        const information = [{heading: `<h1>${content[0].contentTop.heading}</h1>`, text: content[0].contentTop.text}, {heading: `<h3>${content[0].contentBottom.heading}</h3>`, text: content[0].contentBottom.text}];
+        const information = [{
+            heading: `<h1>${content[0].mainHeading}</h1>`, 
+            text: content[0].mainText
+            }, 
+            {
+                heading: `<h3>${content[0].subHeading}</h3>`, 
+                text: content[0].subText
+            }];
 
             return ( 
                 <> 
                     <PageOpener
-                    
+                    information={information} 
+                    image={content[0].image.secure_url}
+                    breadcrumbs={[{to: "/", text: "OUR APPOROACH"}]}
                     >
-                    
-                    
                     
                     <>
                     </>
